// src/parser/godot-resource/types.ts
interface SyntaxNodeBase {
  tree: Tree;
  id: number;
  typeId: number;
  grammarId: number;
  type: string;
  grammarType: string;
  isNamed: boolean;
  isMissing: boolean;
  isExtra: boolean;
  hasChanges: boolean;
  hasError: boolean;
  isError: boolean;
  text: string;
  parseState: number;
  nextParseState: number;
  startPosition: Point;
  endPosition: Point;
  startIndex: number;
  endIndex: number;
  parent: SyntaxNode | null;
  children: Array<SyntaxNode>;
  namedChildren: Array<SyntaxNode>;
  childCount: number;
  namedChildCount: number;
  firstChild: SyntaxNode | null;
  firstNamedChild: SyntaxNode | null;
  lastChild: SyntaxNode | null;
  lastNamedChild: SyntaxNode | null;
  nextSibling: SyntaxNode | null;
  nextNamedSibling: SyntaxNode | null;
  previousSibling: SyntaxNode | null;
  previousNamedSibling: SyntaxNode | null;
  descendantCount: number;

  toString(): string;
  child(index: number): SyntaxNode | null;
  namedChild(index: number): SyntaxNode | null;
  childForFieldName(fieldName: string): SyntaxNode | null;
  childForFieldId(fieldId: number): SyntaxNode | null;
  fieldNameForChild(childIndex: number): string | null;
  childrenForFieldName(fieldName: string): Array<SyntaxNode>;
  childrenForFieldId(fieldId: number): Array<SyntaxNode>;
  firstChildForIndex(index: number): SyntaxNode | null;
  firstNamedChildForIndex(index: number): SyntaxNode | null;

  descendantForIndex(index: number): SyntaxNode;
  descendantForIndex(startIndex: number, endIndex: number): SyntaxNode;
  namedDescendantForIndex(index: number): SyntaxNode;
  namedDescendantForIndex(startIndex: number, endIndex: number): SyntaxNode;
  descendantForPosition(position: Point): SyntaxNode;
  descendantForPosition(startPosition: Point, endPosition: Point): SyntaxNode;
  namedDescendantForPosition(position: Point): SyntaxNode;
  namedDescendantForPosition(startPosition: Point, endPosition: Point): SyntaxNode;
  descendantsOfType<T extends TypeString>(types: T | readonly T[], startPosition?: Point, endPosition?: Point): NodeOfType<T>[];

  closest<T extends SyntaxType>(types: T | readonly T[]): NamedNode<T> | null;
  walk(): TreeCursor;
}
// ...
interface NamedNodeBase extends SyntaxNodeBase {
    isNamed: true;
}

/** An unnamed node with the given type string. */
export interface UnnamedNode<T extends string = string> extends SyntaxNodeBase {
  type: T;
  isNamed: false;
}
// ...
export interface ErrorNode extends NamedNodeBase {
    type: typeof SyntaxType.ERROR;
    hasError: true;
}

export const SyntaxType = {
  ERROR: "ERROR",
  Arguments: "arguments",
  Array: "array",
  Attribute: "attribute",
  Constructor: "constructor",
  Dictionary: "dictionary",
  Identifier: "identifier",
  Pair: "pair",
  Property: "property",
  Resource: "resource",
  Section: "section",
  Comment: "comment",
  False: "false",
  Float: "float",
  Integer: "integer",
  Null: "null",
  Path: "path",
  String: "string",
  True: "true",
} as const;
export type SyntaxType = typeof SyntaxType[keyof typeof SyntaxType];

export type UnnamedType =
  | "("
  | ")"
  | ","
  | ":"
  | "="
  | "["
  | "]"
  | "{"
  | "}"
  ;

export type TypeString = SyntaxType | UnnamedType;

export type SyntaxNode =
  | ArgumentsNode
  | ArrayNode
  | AttributeNode
  | ConstructorNode
  | DictionaryNode
  | IdentifierNode
  | PairNode
  | PropertyNode
  | ResourceNode
  | SectionNode
  | UnnamedNode<"(">
  | UnnamedNode<")">
  | UnnamedNode<",">
  | UnnamedNode<":">
  | UnnamedNode<"=">
  | UnnamedNode<"[">
  | UnnamedNode<"]">
  | CommentNode
  | FalseNode
  | FloatNode
  | IntegerNode
  | NullNode
  | PathNode
  | StringNode
  | TrueNode
  | UnnamedNode<"{">
  | UnnamedNode<"}">
  | ErrorNode
  ;

export interface ArgumentsNode extends NamedNodeBase {
  type: typeof SyntaxType.Arguments;
}

export interface ArrayNode extends NamedNodeBase {
  type: typeof SyntaxType.Array;
}

export interface AttributeNode extends NamedNodeBase {
  type: typeof SyntaxType.Attribute;
}

export interface ConstructorNode extends NamedNodeBase {
  type: typeof SyntaxType.Constructor;
}

export interface DictionaryNode extends NamedNodeBase {
  type: typeof SyntaxType.Dictionary;
}

export interface IdentifierNode extends NamedNodeBase {
  type: typeof SyntaxType.Identifier;
}

export interface PairNode extends NamedNodeBase {
  type: typeof SyntaxType.Pair;
}

export interface PropertyNode extends NamedNodeBase {
  type: typeof SyntaxType.Property;
}

export interface ResourceNode extends NamedNodeBase {
  type: typeof SyntaxType.Resource;
}

export interface SectionNode extends NamedNodeBase {
  type: typeof SyntaxType.Section;
}

export interface CommentNode extends NamedNodeBase {
  type: typeof SyntaxType.Comment;
}

export interface FalseNode extends NamedNodeBase {
  type: typeof SyntaxType.False;
}

export interface FloatNode extends NamedNodeBase {
  type: typeof SyntaxType.Float;
}

export interface IntegerNode extends NamedNodeBase {
  type: typeof SyntaxType.Integer;
}

export interface NullNode extends NamedNodeBase {
  type: typeof SyntaxType.Null;
}

export interface PathNode extends NamedNodeBase {
  type: typeof SyntaxType.Path;
}

export interface StringNode extends NamedNodeBase {
  type: typeof SyntaxType.String;
}

export interface TrueNode extends NamedNodeBase {
  type: typeof SyntaxType.True;
}